Configuration file location is now mandatory.
 * Removed the hardcoded, default configuration file location.
 * The config file must now be specified at runtime as a command line
   argument.
 * Added earlier validation for the port number.
 * Also tweaked File Handler function documentation

+++ server/src/OASServerInfo.cpp
@@ -8,7 +8,7 @@ ServerInfo::ServerInfo()
 }
 
 ServerInfo::ServerInfo( std::string const& cacheDirectory, 
-                        unsigned short port):
+                        long int port):
                         _cacheDirectory(cacheDirectory),
                         _port(port)
 {
@@ -20,7 +20,7 @@ std::string const& ServerInfo::getCacheDirectory() const
     return this->_cacheDirectory;
 }
 
-unsigned short ServerInfo::getPort() const
+long int ServerInfo::getPort() const
 {
     return this->_port;
 }
